Refactor tests to use new record mutations and navigation hooks
