+++ src/test/__mocks__/features/resources/hooks/useRecordMutations.mock.ts
@@ -1,0 +1,7 @@
+export const saveRecord = jest.fn();
+export const deleteRecord = jest.fn();
+export const changeRecordProfile = jest.fn();
+
+jest.mock('@/features/resources/hooks/useRecordMutations', () => ({
+  useRecordMutations: () => ({ saveRecord, deleteRecord, changeRecordProfile }),
+}));
