+++ src/test/__mocks__/features/resources/hooks/useRecordNavigation.mock.ts
@@ -1,0 +1,7 @@
+export const discardRecord = jest.fn();
+export const tryFetchExternalRecordForEdit = jest.fn();
+export const clearRecordState = jest.fn();
+
+jest.mock('@/features/resources/hooks/useRecordNavigation', () => ({
+  useRecordNavigation: () => ({ discardRecord, tryFetchExternalRecordForEdit, clearRecordState }),
+}));
